requirements.txt
# Core scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Machine learning
scikit-learn>=1.3.0

# Bioinformatics
biopython>=1.81

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Model persistence
joblib>=1.3.0

# Optional: Progress bars
tqdm>=4.65.0
